requirements.txt
alembic==1.4.3
aniso8601==9.0.1
appdirs==1.4.4
asgiref==3.2.10
attrs==21.2.0
bcrypt==3.2.0
cachelib==0.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.1.1
distlib==0.3.1
Django==3.1
django-filter==2.3.0
djangorestframework==3.11.1
dnspython==2.2.1
filelock==3.0.12
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-PyMongo==2.3.0
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-Seeder==1.1.1
Flask-Session==0.3.2
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==4.3.2
Mako==1.1.3
Markdown==3.2.2
MarkupSafe==1.1.1
paramiko==2.7.2
pbr==5.5.0
Pillow==7.2.0
psycopg2-binary==2.8.6
pycparser==2.20
pykerberos==1.2.4
pymongo==4.1.1
PyNaCl==1.4.0
pyrsistent==0.18.0
pysftp==0.2.9
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
python-snappy==0.6.1
pytz==2020.1
requests==2.24.0
six==1.15.0
spotipy==2.16.0
SQLAlchemy==1.3.19
sqlparse==0.3.1
stevedore==3.2.2
urllib3==1.25.10
virtualenv==20.0.31
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
Werkzeug==1.0.1
